More CR changes. Moved input handler inspectors into their own folder and namespace.
