+++ Assets/MixedRealityToolkit-SDK/Inspectors/Input/Handlers.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 97dc390099f55d74bae7d4c2938bdb09
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
